move randomToken to common/algo and call it directly from site.sri assignment
